Refactor hoc to hook
