+++ src/pages/index.tsx
@@ -1,13 +1,18 @@
 import type { NextPage } from 'next';
 
-import withUser from 'hocs/withUser';
+import PageLoader from 'components/PageLoader';
+import useSession from 'hooks/useSession';
 
 const Home: NextPage = () => {
+  const [, isUserLoading] = useSession();
+
   return (
-    <h1 className="text-white text-2xl font-bold" role="heading">
-      Coming soon. Stay tune for more content 😎
-    </h1>
+    <PageLoader isLoading={isUserLoading}>
+      <h1 className="text-white text-2xl font-bold" role="heading">
+        Coming soon. Stay tune for more content 😎
+      </h1>
+    </PageLoader>
   );
 };
 
-export default withUser(Home);
+export default Home;
